:bug: test crash fix
