+++ CMakeLists.txt
@@ -48,7 +48,7 @@ add_executable(${PROJECT_NAME}-googletest
 	src/utils/error.cpp
 	
 	# test files
-  test/global_test_env.cpp
+  test/gtest.cpp
 	test/server/TimerTree_test.cpp
   test/server/ConfigHandler/allowRequest/allowRequest.cpp
   test/server/ConfigHandler/allowRequest/limitExcept.cpp
@@ -104,7 +104,7 @@ add_executable(${PROJECT_NAME}-googletest
 	test/cgi/CGIExecutor_test.cpp
 	test/cgi/CGIParser_test.cpp
 )
-target_link_libraries(${PROJECT_NAME}-googletest gtest_main)
+# target_link_libraries(${PROJECT_NAME}-googletest gtest_main)
 # debug
 # target_link_libraries(${PROJECT_NAME}-googletest gtest_main -fsanitize=address -g)
 # target_compile_options(${PROJECT_NAME}-googletest PRIVATE -fsanitize=address -g)
@@ -113,6 +113,7 @@ target_include_directories(${PROJECT_NAME}-googletest PRIVATE
 	src/server
 	src/config
 	src/utils
+  test
 	test/conf
   test/server/ConfigHandler
 	test/server/HttpResponse
