+++ src/main.cpp
@@ -23,5 +23,6 @@ int main(int ac, char *av[]) {
   } catch (const std::exception &e) {
     std::cerr << e.what() << std::endl;
   }
+  delete config;
   return 1;
 }
