+++ test/GlobalTestEnvironment.hpp
@@ -1,4 +1,5 @@
-
+#ifndef GLOBAL_TEST_ENVIRONMENT_H
+#define GLOBAL_TEST_ENVIRONMENT_H
 #include <gtest/gtest.h>
 
 #include "WebServer.hpp"
@@ -19,15 +20,6 @@ class GlobalTestEnvironment : public ::testing::Environment {
 };
 
 // グローバルテスト環境のインスタンスを生成
-::testing::Environment* createGlobalTestEnvironment() {
-  static GlobalTestEnvironment* env = new GlobalTestEnvironment();
-  return env;
-}
+::testing::Environment* createGlobalTestEnvironment();
 
-int main(int argc, char** argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  // グローバルなセットアップとクリーンアップを登録
-  ::testing::AddGlobalTestEnvironment(createGlobalTestEnvironment());
-
-  return RUN_ALL_TESTS();
-}
+#endif
